Ttl fix

+++ app/config/operation.go
@@ -2,4 +2,4 @@ package config
 
 import "time"
 
-var DEF_SESSION_TTL = int(time.Hour * 24 * 7)
+var DEF_SESSION_TTL = time.Hour * 24 * 7
